Maintain version history downstream
Add a .releases.xml file which contains the <release> entries from the
metainfo that used to be kept upstream.

This doesn't affect the current build (since nothing refers to it) but
the next release of cockpit will use this version information to fill
the releases section of its metainfo at install time.

Maintaining this information downstream allows us to simplify the
upstream release process and produces a more coherent version history
for users of the flatpak.